requirements_no_compile.txt
customtkinter==5.2.2
requests==2.32.5
aiohttp==3.9.1
redis==5.0.1
pybreaker==1.4.1
prometheus-client==0.21.0
jaeger-client==4.8.0
twilio==9.3.2
Pillow==12.1.1
qrcode==8.2
openpyxl==3.1.5
pandas==3.0.1
python-dotenv==1.2.2
scikit-learn==1.8.0
numpy==2.4.3
PySocks==1.7.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0